Get post on ticketed class working
