+++ expo/gbe/templates/gbe/scheduling/event_booking.tmpl
@@ -9,7 +9,7 @@
     <p style= "color:red"> There is an error on the form.</p>
   {% endif %}
   <font color="red">
-  {{ form.non_field_errors }}
+  {{ form.non_field_errors }}{{ form.errors }}
   {{ scheduling_form.non_field_errors }}</font>
 <div class="event_booking">
 <h4>Event Details</h4>
